move compressed toolkit to separate file, fix core tests
