+++ src/lib.rs
@@ -3,4 +3,8 @@
 
 pub mod shared_display_ref;
 pub mod toolkit;
+pub mod toolkit_compressed;
+
 pub use shared_display_core::*;
+pub use toolkit::*;
+pub use toolkit_compressed::*;
